// packages/core/src/utils/clients/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      Workflow: {
        Row: {
          created_at: string
          description: string
          updated_at: string
          wf_id: string
        }
        Insert: {
          created_at?: string
          description: string
          updated_at?: string
          wf_id?: string
        }
        Update: {
          created_at?: string
          description?: string
          updated_at?: string
          wf_id?: string
        }
        Relationships: []
      }
      WorkflowInvocation: {
        Row: {
          accuracy: number | null
          actual_output: string | null
          end_time: string | null
          evaluation_inputs: Json | null
          expected_output: string | null
          expected_output_type: Json | null
          extras: Json | null
          feedback: string | null
          fitness: Json | null
          fitness_score: number | null
          generation_id: string | null
          metadata: Json | null
          novelty: number | null
          preparation: string | null
          run_id: string | null
          start_time: string
          status: Database["public"]["Enums"]["InvocationStatus"]
          usd_cost: number
          wf_invocation_id: string
          wf_version_id: string
          workflow_input: Json | null
          workflow_io: Json | null
          workflow_output: Json | null
        }
        Insert: {
          accuracy?: number | null
          actual_output?: string | null
          end_time?: string | null
          evaluation_inputs?: Json | null
          expected_output?: string | null
          expected_output_type?: Json | null
          extras?: Json | null
          feedback?: string | null
          fitness?: Json | null
          fitness_score?: number | null
          generation_id?: string | null
          metadata?: Json | null
          novelty?: number | null
          preparation?: string | null
          run_id?: string | null
          start_time?: string
          status?: Database["public"]["Enums"]["InvocationStatus"]
          usd_cost?: number
          wf_invocation_id?: string
          wf_version_id: string
          workflow_input?: Json | null
          workflow_io?: Json | null
          workflow_output?: Json | null
        }
        Update: {
          accuracy?: number | null
          actual_output?: string | null
          end_time?: string | null
          evaluation_inputs?: Json | null
          expected_output?: string | null
          expected_output_type?: Json | null
          extras?: Json | null
          feedback?: string | null
          fitness?: Json | null
          fitness_score?: number | null
          generation_id?: string | null
          metadata?: Json | null
          novelty?: number | null
          preparation?: string | null
          run_id?: string | null
          start_time?: string
          status?: Database["public"]["Enums"]["InvocationStatus"]
          usd_cost?: number
          wf_invocation_id?: string
          wf_version_id?: string
          workflow_input?: Json | null
          workflow_io?: Json | null
          workflow_output?: Json | null
        }
        Relationships: [
          {
            foreignKeyName: "fk_wfi_generation"
            columns: ["generation_id"]
            isOneToOne: false
            referencedRelation: "Generation"
            referencedColumns: ["generation_id"]
          },
          {
            foreignKeyName: "fk_wfi_run"
            columns: ["run_id"]
            isOneToOne: false
            referencedRelation: "EvolutionRun"
            referencedColumns: ["run_id"]
          },
          {
            foreignKeyName: "WorkflowInvocation_wf_version_id_fkey"
            columns: ["wf_version_id"]
            isOneToOne: false
            referencedRelation: "WorkflowVersion"
            referencedColumns: ["wf_version_id"]
          },
        ]
      }
      WorkflowVersion: {
        Row: {
          all_workflow_io: Json[] | null
          commit_message: string
          created_at: string
          dsl: Json
          generation_id: string | null
          iteration_budget: number
          knowledge: Json | null
          operation: Database["public"]["Enums"]["WorkflowOperation"]
          parent_id: string | null
          parent1_id: string | null
          parent2_id: string | null
          time_budget_seconds: number
          updated_at: string
          wf_version_id: string
          workflow_id: string
        }
        Insert: {
          all_workflow_io?: Json[] | null
          commit_message: string
          created_at?: string
          dsl: Json
          generation_id?: string | null
          iteration_budget?: number
          knowledge?: Json | null
          operation?: Database["public"]["Enums"]["WorkflowOperation"]
          parent_id?: string | null
          parent1_id?: string | null
          parent2_id?: string | null
          time_budget_seconds?: number
          updated_at?: string
          wf_version_id?: string
          workflow_id: string
        }
        Update: {
          all_workflow_io?: Json[] | null
          commit_message?: string
          created_at?: string
          dsl?: Json
          generation_id?: string | null
          iteration_budget?: number
          knowledge?: Json | null
          operation?: Database["public"]["Enums"]["WorkflowOperation"]
          parent_id?: string | null
          parent1_id?: string | null
          parent2_id?: string | null
          time_budget_seconds?: number
          updated_at?: string
          wf_version_id?: string
          workflow_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "fk_wfv_generation"
            columns: ["generation_id"]
            isOneToOne: false
            referencedRelation: "Generation"
            referencedColumns: ["generation_id"]
          },
          {
            foreignKeyName: "fk_wfv_parent1"
            columns: ["parent1_id"]
            isOneToOne: false
            referencedRelation: "WorkflowVersion"
            referencedColumns: ["wf_version_id"]
          },
          {
            foreignKeyName: "fk_wfv_parent2"
            columns: ["parent2_id"]
            isOneToOne: false
            referencedRelation: "WorkflowVersion"
            referencedColumns: ["wf_version_id"]
          },
          {
            foreignKeyName: "fk_workflow_version_parent"
            columns: ["parent_id"]
            isOneToOne: false
            referencedRelation: "WorkflowVersion"
            referencedColumns: ["wf_version_id"]
          },
          {
            foreignKeyName: "WorkflowVersion_workflow_id_fkey"
            columns: ["workflow_id"]
            isOneToOne: false
            referencedRelation: "Workflow"
            referencedColumns: ["wf_id"]
          },
        ]
      }
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never

export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U
      }
      ? U
      : never
    : never